W-10848628 Fix: Added rule to set feature flag for test (#11451)
* W-10848628 Fix: Added rule to set feature flag for test

* Delete extra whitespace

* Fix

(cherry picked from commit 48885f18f3691a48b0a12dcaf7696ca139ea3f07)
